fix lint and cleanup
