+++ packages/grid/_modules_/grid/hooks/utils/useOptionsProp.ts
@@ -4,7 +4,7 @@ import { DEFAULT_LOCALE_TEXT } from '../../constants/localeTextConstants';
 import { GridComponentProps, GridOptionsProp } from '../../GridComponentProps';
 import { ApiRef } from '../../models/api/apiRef';
 import { DEFAULT_GRID_OPTIONS, GridOptions } from '../../models/gridOptions';
-import { getScrollbarSize } from '../../utils/domUtils';
+import { getScrollbarSize } from '../../utils/material-ui-utils';
 import { mergeOptions } from '../../utils/mergeUtils';
 import { useGridReducer } from '../features/core/useGridReducer';
 import { useLogger } from './useLogger';
