+++ packages/grid/_modules_/grid/utils/domUtils.ts
@@ -74,21 +74,3 @@ export function getCellElementFromIndexes(
     `:scope .${CELL_CSS_CLASS}[aria-colIndex='${colIndex}'][data-rowIndex='${rowIndex}']`,
   ) as HTMLDivElement;
 }
-
-// TODO replace with { unstable_ getScrollbarSize } from '@material-ui/utils'
-// A change of the browser zoom change the scrollbar size.
-// Credit https://github.com/twbs/bootstrap/blob/3ffe3a5d82f6f561b82ff78d82b32a7d14aed558/js/src/modal.js#L512-L519
-export function getScrollbarSize(doc: Document): number {
-  const scrollDiv = doc.createElement('div');
-  scrollDiv.style.width = '99px';
-  scrollDiv.style.height = '99px';
-  scrollDiv.style.position = 'absolute';
-  scrollDiv.style.top = '-9999px';
-  scrollDiv.style.overflow = 'scroll';
-
-  doc.body.appendChild(scrollDiv);
-  const scrollbarSize = scrollDiv.offsetWidth - scrollDiv.clientWidth;
-  doc.body.removeChild(scrollDiv);
-
-  return scrollbarSize;
-}
